feat : query donations and profile info from firebase
